+++ lib/onlyoffice_iredmail_helper/mailboxes_methods.rb
@@ -6,7 +6,6 @@ module OnlyofficeIredmailHelper
     # @return [Array<String>] list of folder names
     def mailboxes
       login
-      @imap.select('INBOX')
       folders = @imap.list('%', '%').map(&:name)
       close
       OnlyofficeLoggerHelper.log("Get list of mailboxes: #{folders}")
@@ -18,7 +17,6 @@ module OnlyofficeIredmailHelper
     # @return [nil]
     def create_mailbox(name)
       login
-      @imap.select('INBOX')
       @imap.create(name)
       close
       OnlyofficeLoggerHelper.log("Created new mailbox: #{name}")
@@ -31,10 +29,22 @@ module OnlyofficeIredmailHelper
       raise("There is no mailbox #{name} to delete") unless mailboxes.include?(name)
 
       login
-      @imap.select('INBOX')
       @imap.delete(name)
       close
       OnlyofficeLoggerHelper.log("Delete mailbox by name: #{name}")
+    end
+
+    # Select mailbox with name
+    # @param name [String] name of folder
+    # @return [nil]
+    def mailbox_select(name)
+      login
+
+      return if @current_mailbox == name
+
+      @imap.select(name)
+      @current_mailbox = name
+      OnlyofficeLoggerHelper.log("Selected current mailbox: #{name}")
     end
   end
 end