+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add `yamllint` check in CI
 * Add `ruby-3.2` to CI
+* New `IredMailHelper#mailbox_select` method
 
 ### Fixes
 
@@ -20,6 +21,7 @@
 * Set `--fail-fast` for rspec tests
 * Drop support of `ruby-2.6`, `ruby-2.7`, since it's EOL'ed
 * Extract creating SMTP connection to separate method
+* Force `INBOX` mailbox select on login
 
 ## 1.0.0 (2022-01-27)